% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itksnap.R
\name{itksnap}
\alias{itksnap}
\title{Run ITK-SNAP}
\usage{
itksnap(files, segmentation = NULL, labels = NULL, zoom = NULL,
  verbose = TRUE, ...)
}
\arguments{
\item{files}{filenames or nifti objects to be displayed.
If multiple files are given assumed the first to load and the rest are overlays}

\item{segmentation}{segmentation image using \code{-s} option}

\item{labels}{label descriptions using \code{-s} option}

\item{zoom}{initial zoom in screen pixels/mm}

\item{verbose}{Print out the command executed}

\item{...}{arguments to pass to \code{\link{itksnap_cmd}}}
}
\value{
Return of \code{\link{system}} command
}
\description{
Executes ITK-SNAP
}

