% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itksnap_cmd.R
\name{itksnap_cmd}
\alias{itksnap_cmd}
\title{Get ITK-SNAP command}
\usage{
itksnap_cmd(arch = R.Version()$arch)
}
\arguments{
\item{arch}{architecture used (i686 or x86_64}
}
\value{
Path of ITK-SNAP
}
\description{
Gets ITK-SNAP command
}

